+++ lib/interface.ts
@@ -45,10 +45,18 @@ export interface IOptions extends IOptionsBase {
     onData?: (data: any) => void;
     /** 上传文件进度 */
     onProgress?: (e?: ProgressEvent) => void;
+    /** 自定义响应数据 */
+    transformData?: (
+        /** 数据 */
+        // eslint-disable-next-line @typescript-eslint/no-explicit-any
+        data: any,
+        /** 响应头 */
+        // eslint-disable-next-line @typescript-eslint/no-explicit-any
+        headers?: { [name: string]: any }
+        // eslint-disable-next-line @typescript-eslint/no-explicit-any
+    ) => any;
     /** 自定义选项，用来传递值自定义处理逻辑 */
     // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    /** 自定义响应体 */
-    transformResponse?: (response: IResult, responseHeaders?: { [name: string]: any }) => IResult;
     [name: string]: any;
 }